create new firebase model, replace withAddWalletConnector with withInitNewWalletConnector, move part of the model/walletconnect to redux/walletconnect
