--- src/hoc/withAddWalletConnector.js
@@ -1,4 +1,0 @@
-import { connect } from 'react-redux';
-import { addWalletConnector } from '../redux/walletconnect';
-
-export default Component => connect(null, { addWalletConnector })(Component);
