+++ src/hoc/withInitNewWalletConnector.js
@@ -1,0 +1,7 @@
+import { connect } from 'react-redux';
+import { walletConnectInitNewSession } from '../redux/walletconnect';
+
+export default Component => connect(
+  null,
+  { walletConnectInitNewSession },
+)(Component);
