+++ src/screens/QRScannerScreenWithData.js
@@ -3,64 +3,38 @@ import lang from 'i18n-js';
 import PropTypes from 'prop-types';
 import React, { PureComponent } from 'react';
 import { Vibration } from 'react-native';
-import firebase from 'react-native-firebase';
 import Piwik from 'react-native-matomo';
 import { compose } from 'recompact';
 import { Alert } from '../components/alerts';
-import { walletConnectInit } from '../model/walletconnect';
-import { withAccountAddress, withAddWalletConnector } from '../hoc';
+import { withAccountAddress, withInitNewWalletConnector } from '../hoc';
 import { getEthereumAddressFromQRCodeData } from '../utils';
 import QRScannerScreen from './QRScannerScreen';
 
 class QRScannerScreenWithData extends PureComponent {
   static propTypes = {
     accountAddress: PropTypes.string,
-    addWalletConnector: PropTypes.func,
+    walletConnectInitNewSession: PropTypes.func,
     isScreenActive: PropTypes.bool,
     navigation: PropTypes.object,
     setSafeTimeout: PropTypes.func,
-  }
+  };
 
-  state = { enableScanning: true }
+  state = { enableScanning: true };
 
-  componentDidUpdate = (prevProps) => {
+  componentDidUpdate = prevProps => {
     if (this.props.isScreenActive && !prevProps.isScreenActive) {
       this.setState({ enableScanning: true });
       Piwik.trackScreen('QRScannerScreen', 'QRScannerScreen');
     }
-  }
+  };
 
-  handlePressBackButton = () => this.props.navigation.navigate('WalletScreen')
+  handlePressBackButton = () => this.props.navigation.navigate('WalletScreen');
 
-  handleReenableScanning = () => this.setState({ enableScanning: true })
-
-  checkPushNotificationPermissions = async () => {
-    const arePushNotificationsAuthorized = await firebase
-      .messaging()
-      .hasPermission();
-
-    if (!arePushNotificationsAuthorized) {
-      // TODO: try catch around Alert?
-      Alert({
-        buttons: [{
-          onPress: async () => firebase.messaging().requestPermission(),
-          text: 'Okay',
-        }, {
-          style: 'cancel',
-          text: 'Dismiss',
-        }],
-        message: lang.t('wallet.push_notifications.please_enable_body'),
-        title: lang.t('wallet.push_notifications.please_enable_title'),
-      });
-    }
-  }
+  handleReenableScanning = () => this.setState({ enableScanning: true });
 
   handleScanSuccess = async ({ data }) => {
     const {
-      accountAddress,
-      addWalletConnector,
-      navigation,
-      setSafeTimeout,
+      accountAddress, walletConnectInitNewSession, navigation, setSafeTimeout,
     } = this.props;
 
     if (!data) return null;
@@ -78,9 +52,7 @@ class QRScannerScreenWithData extends PureComponent {
 
     if (data.startsWith('ethereum:wc')) {
       Piwik.trackEvent('QRScanner', 'walletconnect', 'QRScannedWC');
-      const walletConnector = await walletConnectInit(accountAddress, data);
-      await this.checkPushNotificationPermissions();
-      addWalletConnector(walletConnector);
+      await walletConnectInitNewSession(accountAddress, data);
       return setSafeTimeout(this.handleReenableScanning, 1000);
     }
 
@@ -90,7 +62,7 @@ class QRScannerScreenWithData extends PureComponent {
       title: lang.t('wallet.unrecognized_qrcode_title'),
       callback: this.handleReenableScanning,
     });
-  }
+  };
 
   render = () => (
     <QRScannerScreen
@@ -99,11 +71,11 @@ class QRScannerScreenWithData extends PureComponent {
       onPressBackButton={this.handlePressBackButton}
       onScanSuccess={this.handleScanSuccess}
     />
-  )
+  );
 }
 
 export default compose(
   withAccountAddress,
-  withAddWalletConnector,
+  withInitNewWalletConnector,
   withSafeTimeout,
 )(QRScannerScreenWithData);
